Add thumbnail downloading and (clientside) caching

+++ client/src/javascript/models/images.js
@@ -59,13 +59,15 @@ const Images = {
   },
 
   delete: (id, cb) => {
-    Images.get(id, image =>
+    Images.get(id, (image) => {
       // Check file exists & we have write access
-      fs.access(image.location, fs.constants.W_OK, (err) => {
+      if (!image.location) return cb();
+      return fs.access(image.location, fs.constants.W_OK, (err) => {
         if (err) return cb(`Couldn't access ${image.location}: ${err}`);
+        Sync.clearThumbnail(image.remoteId);
         return fs.unlink(image.location, cb);
-      })
-    );
+      });
+    });
   },
 
   update: (id, data, cb) => {
@@ -78,6 +80,7 @@ const Images = {
   remove: (id, cb) => {
     image_db.findOne({ $loki: id }, (doc) => {
       image_db.removeOne({ $loki: id }, (res) => {
+        if (doc.remoteId) Sync.clearThumbnail(doc.remoteId);
         if (doc && doc.remoteId && Host.isAuthed()) {
           Sync.removeSynced(doc.remoteId, (err) => {
             if (err) cb(err);
